Put all environmental stuff to environment.ts, pass Env as parameter to awilix and db
